add crud

+++ api2/blueprints/comments.py
@@ -8,9 +8,9 @@ from api2.models import Post
 from api2.models.comments import CommentSchema, Comment
 
 
-@comment_api.route('/create/onPost/<int:post_id>', methods=['POST'])
+@comment_api.route('/add/onPost/<int:post_id>', methods=['POST'])
 @jwt_required()
-def comment_post(post_id):
+def add_comment(post_id):
     json_data = request.json
     comment_schema = CommentSchema()
 
@@ -27,22 +27,41 @@ def comment_post(post_id):
     return {"message": f"You have just commented the post with {post_id} id."}
 
 
-@comment_api.route('/all/post/<int:post_id>', methods=['GET'])
+@comment_api.route('/all/inPost/<int:post_id>', methods=['GET'])
 @jwt_required()
-def post_comments(post_id):
+def post_comments_list(post_id):
     comments = Comment.query.filter(Comment.post_id == post_id)
     schema = CommentSchema(many=True)
 
     return jsonify(schema.dump(comments))
 
 
-@comment_api.route('/delete/<int:comment_id>/comment/fromPost/<int:post_id>', methods=["DELETE"])
+@comment_api.route('/delete/<int:comment_id>', methods=["DELETE"])
 @jwt_required()
 def delete_comment(comment_id, post_id):
-    post = Post.query.filter(Post.id == post_id).first()
+    post = Post.query.filter(Post.id == post_id).first_or_404()
+
+    if not current_user.has_commented(post):
+        return abort(400, {"message": "You didn't comment this post"})
+
     current_user.delete_comment(post, comment_id)
     db.session.commit()
     return {"message": f"Comment with {comment_id} id was deleted."}
 
 
+@comment_api.route('/update/<int:comment_id>', methods=["PATCH"])
+@jwt_required()
+def update_comment(comment_id):
+    comment = Comment.query.filter(Comment.id == comment_id).first_or_404()
+    comment_schema = CommentSchema()
+    json_data = request.json
 
+    try:
+        data = comment_schema.load(json_data)
+    except ValidationError:
+        return abort(400, description='Invalid data type. Expected "string".')
+
+    comment.query.update(data)
+    db.session.add(comment)
+    db.session.commit()
+    return comment_schema.dump(comment)
